+++ src/index.html
@@ -1,5 +1,6 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
@@ -7,13 +8,10 @@
   <title>Document</title>
   <link rel="stylesheet" href="styles.css" />
 </head>
+
 <body>
- 
- <div class="root" id="root"></div>
-
- <!--<script src="vistas/login.js"></script>-->
+  <div class="root" id="root"></div>
   <script type="module" src="main.js"></script>
-  
 </body>
-</html>
 
+</html>
