+++ src/main.js
@@ -1,38 +1,34 @@
-import {signUp } from "./vistas/SignUp.js"; 
-import {signIn } from "./vistas/login.js"; 
-import {authGoogle} from "./assets/login-assets.js"
-import {signIn } from "./vistas/SignIn.js"; 
-/*import { FirebaseApp } from "../../compilacion";
-import { getAuth, createUserWithEmailAndPassword} from "firebase/auth"
-const auth= getAuth(FirebaseApp)*/
+import { signUp } from "./vistas/SignUp.js";
+import { signIn } from "./vistas/SignIn.js";
+import { firebaseInit, authGoogle } from "./lib/compilacion.js";
 
 //div que contiene formularios en Html
 const root = document.getElementById("root");
 // Iniciar Sesión
-export const iniciar= root.appendChild(signIn());
+export const iniciar = root.appendChild(signIn());
 //Registrar
 let signUpButton = document.getElementById("buttonRegistrar");
 
-  signUpButton.addEventListener('click',() => {
+signUpButton.addEventListener('click', () => {
 
-        root.innerHTML=" "; 
-        root.appendChild(signUp());
-        signUp().style.display = "block";
-        iniciar.style.display = "none";
-let btnSU=document.getElementById('button-SignUp');
-
-//Login with google
-document.getElementById('iniciaGoogle').addEventListener('click',()=>{
-        authGoogle();
+  root.innerHTML = " ";
+  root.appendChild(signUp());
+  signUp().style.display = "block";
+  iniciar.style.display = "none";
+  let btnSU = document.getElementById('button-SignUp');
+  btnSU.addEventListener('click', (e) => {
+    e.preventDefault();
+    let signupEmail = document.getElementById('Email').value;
+    let signupContraseña = document.getElementById('password').value;
+    console.log(signupEmail, signupContraseña);
+  }
+  );
 });
 
- 
-  btnSU.addEventListener('click', (e) => {
-    e.preventDefault();       
-let signupEmail = document.getElementById('Email').value;
-let signupContraseña = document.getElementById('password').value;
-console.log(signupEmail, signupContraseña );
-      }
-      ); 
-  });
+//Login with google
+document.getElementById('iniciaGoogle').addEventListener('click', (e) => {
+  e.preventDefault();
+  authGoogle();
+});
 
+firebaseInit();
