--- src/vistas/logueoregistrar.js
@@ -1,13 +1,0 @@
-//import {registrar } from "./main.js"; 
- 
-
-//const btonsignup = document.querySelectorAll('button-SignUp');
-//console.log(btonsignup);
-
-
- 
-    
-
-
-
-
